Moving services into their own folder
