+++ src/main/java/org/icatproject/ids/dataSelection/DataSelectionFactory.java
@@ -18,8 +18,6 @@ import org.icatproject.IcatExceptionType;
 import org.icatproject.IcatException_Exception;
 import org.icatproject.icat.client.IcatException;
 import org.icatproject.icat.client.Session;
-import org.icatproject.ids.IcatReader;
-import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.enums.RequestType;
 import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.exceptions.BadRequestException;
@@ -31,7 +29,9 @@ import org.icatproject.ids.helpers.LocationHelper;
 import org.icatproject.ids.models.DataFileInfo;
 import org.icatproject.ids.models.DataInfoBase;
 import org.icatproject.ids.models.DataSetInfo;
-import org.icatproject.ids.v3.ServiceProvider;
+import org.icatproject.ids.services.IcatReader;
+import org.icatproject.ids.services.PropertyHandler;
+import org.icatproject.ids.services.ServiceProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
