+++ src/main/java/org/icatproject/ids/dataSelection/DataSelectionForStorageUnitDatafile.java
@@ -9,7 +9,7 @@ import org.icatproject.ids.enums.RequestType;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.exceptions.NotImplementedException;
 import org.icatproject.ids.models.DataInfoBase;
-import org.icatproject.ids.v3.ServiceProvider;
+import org.icatproject.ids.services.ServiceProvider;
 
 public class DataSelectionForStorageUnitDatafile extends DataSelectionBase {
