+++ src/main/java/org/icatproject/ids/finiteStateMachine/FiniteStateMachineForStorageUnitDatafile.java
@@ -15,16 +15,16 @@ import java.util.Map.Entry;
 import java.util.TimerTask;
 
 import org.icatproject.Dataset;
-import org.icatproject.ids.IcatReader;
-import org.icatproject.ids.LockManager;
-import org.icatproject.ids.LockManager.Lock;
-import org.icatproject.ids.LockManager.LockType;
 import org.icatproject.ids.enums.DeferredOp;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.models.DataFileInfo;
 import org.icatproject.ids.models.DataInfoBase;
 import org.icatproject.ids.models.DataSetInfo;
 import org.icatproject.ids.plugin.AlreadyLockedException;
+import org.icatproject.ids.services.IcatReader;
+import org.icatproject.ids.services.LockManager;
+import org.icatproject.ids.services.LockManager.Lock;
+import org.icatproject.ids.services.LockManager.LockType;
 import org.icatproject.ids.thread.DfArchiver;
 import org.icatproject.ids.thread.DfDeleter;
 import org.icatproject.ids.thread.DfRestorer;
