+++ src/main/java/org/icatproject/ids/finiteStateMachine/FiniteStateMachineForStorageUnitDataset.java
@@ -11,15 +11,15 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.TimerTask;
 
-import org.icatproject.ids.IcatReader;
-import org.icatproject.ids.LockManager;
-import org.icatproject.ids.LockManager.Lock;
-import org.icatproject.ids.LockManager.LockType;
 import org.icatproject.ids.enums.DeferredOp;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.models.DataInfoBase;
 import org.icatproject.ids.models.DataSetInfo;
 import org.icatproject.ids.plugin.AlreadyLockedException;
+import org.icatproject.ids.services.IcatReader;
+import org.icatproject.ids.services.LockManager;
+import org.icatproject.ids.services.LockManager.Lock;
+import org.icatproject.ids.services.LockManager.LockType;
 import org.icatproject.ids.thread.DsArchiver;
 import org.icatproject.ids.thread.DsRestorer;
 import org.icatproject.ids.thread.DsWriter;
