+++ src/main/java/org/icatproject/ids/helpers/SO.java
@@ -10,11 +10,11 @@ import java.util.zip.ZipEntry;
 import java.util.zip.ZipException;
 import java.util.zip.ZipOutputStream;
 
-import org.icatproject.ids.LockManager.Lock;
 import org.icatproject.ids.models.DataFileInfo;
 import org.icatproject.ids.models.DataInfoBase;
 import org.icatproject.ids.models.DataSetInfo;
-import org.icatproject.ids.v3.ServiceProvider;
+import org.icatproject.ids.services.ServiceProvider;
+import org.icatproject.ids.services.LockManager.Lock;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
