+++ src/main/java/org/icatproject/ids/requestHandlers/IsPreparedHandler.java
@@ -24,8 +24,7 @@ import org.icatproject.ids.exceptions.NotFoundException;
 import org.icatproject.ids.exceptions.NotImplementedException;
 import org.icatproject.ids.helpers.ValueContainer;
 import org.icatproject.ids.models.Prepared;
-import org.icatproject.ids.v3.RequestHandlerBase;
-import org.icatproject.ids.v3.ServiceProvider;
+import org.icatproject.ids.services.ServiceProvider;
 
 import jakarta.json.Json;
 import jakarta.json.stream.JsonGenerator;
