+++ src/main/java/org/icatproject/ids/requestHandlers/WriteHandler.java
@@ -6,8 +6,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 import org.icatproject.IcatException_Exception;
-import org.icatproject.ids.LockManager.Lock;
-import org.icatproject.ids.LockManager.LockType;
 import org.icatproject.ids.dataSelection.DataSelectionBase;
 import org.icatproject.ids.enums.CallType;
 import org.icatproject.ids.enums.DeferredOp;
@@ -22,8 +20,9 @@ import org.icatproject.ids.exceptions.NotImplementedException;
 import org.icatproject.ids.helpers.ValueContainer;
 import org.icatproject.ids.models.DataInfoBase;
 import org.icatproject.ids.plugin.AlreadyLockedException;
-import org.icatproject.ids.v3.RequestHandlerBase;
-import org.icatproject.ids.v3.ServiceProvider;
+import org.icatproject.ids.services.ServiceProvider;
+import org.icatproject.ids.services.LockManager.Lock;
+import org.icatproject.ids.services.LockManager.LockType;
 
 import jakarta.json.Json;
 import jakarta.json.stream.JsonGenerator;
