+++ src/main/java/org/icatproject/ids/services/ServiceProvider.java
@@ -1,12 +1,8 @@
-package org.icatproject.ids.v3;
+package org.icatproject.ids.services;
 
 import java.util.Set;
 
 import org.icatproject.ICAT;
-import org.icatproject.ids.IcatReader;
-import org.icatproject.ids.LockManager;
-import org.icatproject.ids.PropertyHandler;
-import org.icatproject.ids.Transmitter;
 import org.icatproject.ids.enums.CallType;
 import org.icatproject.ids.finiteStateMachine.FiniteStateMachine;
 import org.icatproject.ids.plugin.MainStorageInterface;
