+++ src/main/java/org/icatproject/ids/thread/DsArchiver.java
@@ -5,11 +5,11 @@ import java.nio.file.Path;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.icatproject.ids.LockManager.Lock;
 import org.icatproject.ids.finiteStateMachine.FiniteStateMachine;
 import org.icatproject.ids.models.DataSetInfo;
-import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.plugin.MainStorageInterface;
+import org.icatproject.ids.services.PropertyHandler;
+import org.icatproject.ids.services.LockManager.Lock;
 
 /*
  * Removes datasets from the fast storage (doesn't write them to slow storage)
