+++ src/main/java/org/icatproject/ids/thread/DsWriter.java
@@ -14,16 +14,16 @@ import org.slf4j.LoggerFactory;
 
 import org.icatproject.Datafile;
 import org.icatproject.Dataset;
-import org.icatproject.ids.IcatReader;
-import org.icatproject.ids.LockManager.Lock;
 import org.icatproject.ids.finiteStateMachine.FiniteStateMachine;
 import org.icatproject.ids.helpers.LocationHelper;
 import org.icatproject.ids.models.DataFileInfo;
 import org.icatproject.ids.models.DataSetInfo;
-import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.plugin.ArchiveStorageInterface;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.plugin.ZipMapperInterface;
+import org.icatproject.ids.services.IcatReader;
+import org.icatproject.ids.services.PropertyHandler;
+import org.icatproject.ids.services.LockManager.Lock;
 
 /**
  * Copies dataset from main to archive
